feat(fonts): trying to fix the oxanium bug
