+++ packages/API/.eslintcache
@@ -1,1 +1,1 @@
-[{"/workspaces/Athenetz-SV/packages/API/src/firebase/firebase.ts":"1","/workspaces/Athenetz-SV/packages/API/src/other/gh/GitAvatar.ts":"2"},{"size":1956,"mtime":1727438876236,"results":"3","hashOfConfig":"4"},{"size":651,"mtime":1727438876236,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","suppressedMessages":"8","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"g68oho",{"filePath":"9","messages":"10","suppressedMessages":"11","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/workspaces/Athenetz-SV/packages/API/src/firebase/firebase.ts",[],["12","13","14","15","16","17","18","19","20","21","22","23","24"],"/workspaces/Athenetz-SV/packages/API/src/other/gh/GitAvatar.ts",[],["25"],{"ruleId":"26","severity":2,"message":"27","line":10,"column":11,"nodeType":"28","messageId":"29","endLine":10,"endColumn":18,"suppressions":"30"},{"ruleId":"26","severity":2,"message":"27","line":11,"column":15,"nodeType":"28","messageId":"29","endLine":11,"endColumn":22,"suppressions":"31"},{"ruleId":"26","severity":2,"message":"27","line":12,"column":14,"nodeType":"28","messageId":"29","endLine":12,"endColumn":21,"suppressions":"32"},{"ruleId":"26","severity":2,"message":"27","line":13,"column":18,"nodeType":"28","messageId":"29","endLine":13,"endColumn":25,"suppressions":"33"},{"ruleId":"26","severity":2,"message":"27","line":14,"column":22,"nodeType":"28","messageId":"29","endLine":14,"endColumn":29,"suppressions":"34"},{"ruleId":"26","severity":2,"message":"27","line":15,"column":10,"nodeType":"28","messageId":"29","endLine":15,"endColumn":17,"suppressions":"35"},{"ruleId":"26","severity":2,"message":"27","line":16,"column":18,"nodeType":"28","messageId":"29","endLine":16,"endColumn":25,"suppressions":"36"},{"ruleId":"26","severity":2,"message":"37","line":21,"column":3,"nodeType":"28","messageId":"29","endLine":21,"endColumn":13,"suppressions":"38"},{"ruleId":"26","severity":2,"message":"37","line":29,"column":3,"nodeType":"28","messageId":"29","endLine":29,"endColumn":13,"suppressions":"39"},{"ruleId":"26","severity":2,"message":"27","line":37,"column":3,"nodeType":"28","messageId":"29","endLine":37,"endColumn":10,"suppressions":"40"},{"ruleId":"41","severity":2,"message":"42","line":41,"column":11,"nodeType":"28","messageId":"43","endLine":41,"endColumn":21,"suppressions":"44"},{"ruleId":"45","severity":2,"message":"46","line":41,"column":24,"nodeType":"47","endLine":41,"endColumn":75,"suppressions":"48"},{"ruleId":"26","severity":2,"message":"27","line":41,"column":24,"nodeType":"28","messageId":"29","endLine":41,"endColumn":31,"suppressions":"49"},{"ruleId":"26","severity":2,"message":"27","line":7,"column":31,"nodeType":"28","messageId":"29","endLine":7,"endColumn":38,"suppressions":"50"},"no-undef","'process' is not defined.","Identifier","undef",["51"],["52"],["53"],["54"],["55"],["56"],["57"],"'globalThis' is not defined.",["58"],["59"],["60"],"no-unused-vars","'recv3token' is assigned a value but never used.","unusedVar",["61"],"turbo/no-undeclared-env-vars","NEXT_PUBLIC_GOOGLE_RECAPTCHA_PUBLIC_KEY is not listed as a dependency in turbo.json","MemberExpression",["62"],["63"],["64"],{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},"directive",""]
+[{"D:\\dev\\Athenetz-SV\\packages\\API\\src\\firebase\\firebase.ts":"1","D:\\dev\\Athenetz-SV\\packages\\API\\src\\other\\gh\\GitAvatar.ts":"2"},{"size":2021,"mtime":1727465256822,"results":"3","hashOfConfig":"4"},{"size":651,"mtime":1727283111142,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","suppressedMessages":"8","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1ir6frf",{"filePath":"9","messages":"10","suppressedMessages":"11","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\dev\\Athenetz-SV\\packages\\API\\src\\firebase\\firebase.ts",[],["12","13","14","15","16","17","18","19","20","21","22","23","24"],"D:\\dev\\Athenetz-SV\\packages\\API\\src\\other\\gh\\GitAvatar.ts",[],["25"],{"ruleId":"26","severity":2,"message":"27","line":10,"column":11,"nodeType":"28","messageId":"29","endLine":10,"endColumn":18,"suppressions":"30"},{"ruleId":"26","severity":2,"message":"27","line":11,"column":15,"nodeType":"28","messageId":"29","endLine":11,"endColumn":22,"suppressions":"31"},{"ruleId":"26","severity":2,"message":"27","line":12,"column":14,"nodeType":"28","messageId":"29","endLine":12,"endColumn":21,"suppressions":"32"},{"ruleId":"26","severity":2,"message":"27","line":13,"column":18,"nodeType":"28","messageId":"29","endLine":13,"endColumn":25,"suppressions":"33"},{"ruleId":"26","severity":2,"message":"27","line":14,"column":22,"nodeType":"28","messageId":"29","endLine":14,"endColumn":29,"suppressions":"34"},{"ruleId":"26","severity":2,"message":"27","line":15,"column":10,"nodeType":"28","messageId":"29","endLine":15,"endColumn":17,"suppressions":"35"},{"ruleId":"26","severity":2,"message":"27","line":16,"column":18,"nodeType":"28","messageId":"29","endLine":16,"endColumn":25,"suppressions":"36"},{"ruleId":"26","severity":2,"message":"37","line":21,"column":3,"nodeType":"28","messageId":"29","endLine":21,"endColumn":13,"suppressions":"38"},{"ruleId":"26","severity":2,"message":"37","line":29,"column":3,"nodeType":"28","messageId":"29","endLine":29,"endColumn":13,"suppressions":"39"},{"ruleId":"26","severity":2,"message":"27","line":37,"column":3,"nodeType":"28","messageId":"29","endLine":37,"endColumn":10,"suppressions":"40"},{"ruleId":"41","severity":2,"message":"42","line":41,"column":11,"nodeType":"28","messageId":"43","endLine":41,"endColumn":21,"suppressions":"44"},{"ruleId":"45","severity":2,"message":"46","line":41,"column":24,"nodeType":"47","endLine":41,"endColumn":75,"suppressions":"48"},{"ruleId":"26","severity":2,"message":"27","line":41,"column":24,"nodeType":"28","messageId":"29","endLine":41,"endColumn":31,"suppressions":"49"},{"ruleId":"26","severity":2,"message":"27","line":7,"column":31,"nodeType":"28","messageId":"29","endLine":7,"endColumn":38,"suppressions":"50"},"no-undef","'process' is not defined.","Identifier","undef",["51"],["52"],["53"],["54"],["55"],["56"],["57"],"'globalThis' is not defined.",["58"],["59"],["60"],"no-unused-vars","'recv3token' is assigned a value but never used.","unusedVar",["61"],"turbo/no-undeclared-env-vars","NEXT_PUBLIC_GOOGLE_RECAPTCHA_PUBLIC_KEY is not listed as a dependency in turbo.json","MemberExpression",["62"],["63"],["64"],{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},{"kind":"65","justification":"66"},"directive",""]
