+++ packages/SV-UI/src/fonts.ts
@@ -1,0 +1,15 @@
+import { Oxanium, Bungee_Spice } from "next/font/google";
+
+
+const oxanium = Oxanium({
+  subsets: ["latin"],
+  display: "swap",
+});
+const bungeeSpice = Bungee_Spice({
+  weight: "400",
+  subsets: ["latin"],
+  display: "swap",
+});
+
+
+export { oxanium, bungeeSpice };
