+++ packages/SV-UI/src/presets/className.tsx
@@ -1,19 +1,7 @@
-import { Oxanium, Bungee_Spice } from "next/font/google";
-const oxanium = Oxanium({
-  subsets: ["latin"],
-  display: "swap",
-  variable: "--font-oxanium",
-});
+import * as fonts from "../fonts";
 
-const bungeeSpice = Bungee_Spice({
-  weight: "400",
-  subsets: ["latin"],
-  display: "swap",
-  variable: "--font-bungee-spice",
-});
+export const DefaultWrapper: string = `${fonts.oxanium.className} bg-l-bg dark:bg-d-bg text-l-txt dark:text-d-txt duration-700`;
 
-export const DefaultWrapper: string = `${oxanium.variable} ${bungeeSpice.variable} font-oxanium bg-l-bg dark:bg-d-bg text-l-txt dark:text-d-txt duration-700`;
+export const DefaultText: string = `${fonts.oxanium.className} text-l-txt dark:text-d-txt duration-700`;
 
-export const DefaultText: string = `font-oxanium text-l-txt dark:text-d-txt duration-700`;
-
-export const DefaultHeader: string = `font-bungee-spice text-l-txt dark:text-d-txt duration-700 mb-4 text-2xl font-bold`;
+export const DefaultHeader: string = `${fonts.bungeeSpice.className} text-l-txt dark:text-d-txt duration-700 mb-4 text-2xl font-bold`;
